Simplify calls to makedie
