+++ lib/common.js
@@ -274,3 +274,128 @@ exports.print = function print (err, out) {
     console.dir(arguments[i])
   }
 }
+
+
+exports.makedie = function (instance, ctxt) {
+  ctxt = _.extend(ctxt, instance.die ? instance.die.context : {})
+
+  var die = function (err) {
+    var die_trace = '\n' + (new Error('die trace').stack)
+        .match(/^.*?\n.*\n(.*)/)[1]
+
+    try {
+      if (!err) {
+        err = new Error('unknown')
+      }
+      else if (!Util.isError(err)) {
+        err = new Error(_.isString(err) ? err : Util.inspect(err))
+      }
+
+      err.fatal$ = true
+
+      var so = instance.options()
+
+      // undead is only for testing, do not use in production
+      var undead = (so.debug && so.debug.undead) || (err && err.undead)
+
+      var logargs = [ctxt.type, ctxt.plugin, ctxt.tag, ctxt.id,
+        err.code, err.message, err.details,
+        instance.fixedargs.fatal$ ? 'all-errors-fatal' : '-',
+        ctxt.callpoint()]
+
+      instance.log.fatal.apply(instance, logargs)
+
+      var stack = err.stack || ''
+      stack = stack.replace(/^.*?\n/, '\n')
+
+      var procdesc = '\n  pid=' + process.pid +
+        ', arch=' + process.arch +
+        ', platform=' + process.platform +
+        ',\n  path=' + process.execPath +
+        ',\n  argv=' + Util.inspect(process.argv).replace(/\n/g, '') +
+        ',\n  env=' + Util.inspect(process.env).replace(/\n/g, '')
+
+      var fatalmodemsg = instance.fixedargs.fatal$
+        ? '\n  ALL ERRORS FATAL: action called with argument fatal$:true ' +
+        '(probably a plugin init error, or using a plugin seneca instance' +
+        ', see senecajs.org/fatal.html)' : ''
+
+      var stderrmsg =
+      '\n\n' +
+        'Seneca Fatal Error\n' +
+        '==================\n\n' +
+        'Message: ' + err.message + '\n\n' +
+        'Code: ' + err.code + '\n\n' +
+        'Details: ' + Util.inspect(err.details, {depth: null}) + '\n\n' +
+        'Stack: ' + stack + '\n\n' +
+        'Instance: ' + instance.toString() + fatalmodemsg + die_trace + '\n\n' +
+        'When: ' + new Date().toISOString() + '\n\n' +
+        'Log: ' + Jsonic.stringify(logargs) + '\n\n' +
+        'Node:\n  ' + Util.inspect(process.versions).replace(/\s+/g, ' ') +
+        ',\n  ' + Util.inspect(process.features).replace(/\s+/g, ' ') +
+        ',\n  ' + Util.inspect(process.moduleLoadList).replace(/\s+/g, ' ') + '\n\n' +
+        'Process: ' + procdesc + '\n\n'
+
+      if (so.errhandler) {
+        so.errhandler.call(instance, err)
+      }
+
+      if (instance.closed) {
+        return
+      }
+
+      if (!undead) {
+        instance.act('role:seneca,info:fatal,closing$:true', {err: err})
+
+        instance.close(
+          // terminate process, err (if defined) is from seneca.close
+          function (err) {
+            if (!undead) {
+              process.nextTick(function () {
+                if (err) {
+                  exports.console_error(err)
+                }
+
+                exports.console_error(stderrmsg)
+                exports.console_error('\n\nSENECA TERMINATED at ' + (new Date().toISOString()) +
+                  '. See above for error report.\n\n')
+                process.exit(1)
+              })
+            }
+          }
+       )
+      }
+
+      // make sure we close down within options.deathdelay seconds
+      if (!undead) {
+        var killtimer = setTimeout(function () {
+          exports.console_error(stderrmsg)
+          exports.console_error('\n\nSENECA TERMINATED (on timeout) at ' +
+            (new Date().toISOString()) + '.\n\n')
+          process.exit(2)
+        }, so.deathdelay)
+        killtimer.unref()
+      }
+    }
+    catch (panic) {
+      var msg =
+      '\n\n' +
+        'Seneca Panic\n' +
+        '============\n\n' +
+        panic.stack +
+        '\n\nOriginal Error:\n' +
+        (arguments[0] && arguments[0].stack ? arguments[0].stack : arguments[0])
+      exports.console_error(msg)
+    }
+  }
+
+  die.context = ctxt
+
+  return die
+}
+
+
+// Intentional console errors use this function. Helps to find spurious debugging.
+exports.console_error = function () {
+  console.error.apply(null, arguments)
+}
