+++ lib/plugins.js
@@ -39,7 +39,7 @@ module.exports = function () {
   seneca.decorate('plugins', internals.all)
 }
 
-module.exports.register = function (so, makedie, callpoint) {
+module.exports.register = function (so, callpoint) {
   return function api_register (plugin) {
     var seneca = this
 
@@ -53,7 +53,7 @@ module.exports.register = function (so, makedie, callpoint) {
 
     plugin.fullname = fullname
 
-    var delegate = make_delegate(seneca, plugin, { makedie: makedie })
+    var delegate = make_delegate(seneca, plugin)
 
     seneca.log.debug('register', 'init', fullname, callpoint())
 
@@ -215,7 +215,7 @@ function resolve_options (fullname, plugindef, seneca_options) {
   return outopts
 }
 
-function make_delegate (instance, plugin, options) {
+function make_delegate (instance, plugin) {
   // Adjust Seneca API to be plugin specific.
   var delegate = instance.delegate({
     plugin$: {
@@ -232,7 +232,7 @@ function make_delegate (instance, plugin, options) {
     var args = Common.arrayify(arguments)
 
     // TODO: move outside log function - need handle __update_plugin__
-    var tag = '-' === plugin.tag ? void 0 : plugin.tag
+    var tag = ('-' === plugin.tag) ? void 0 : plugin.tag
     var plugin_fullname = plugin.name + (tag ? '/' + tag : '')
     var plugin_fullname_wide =
           (plugin_fullname.length < 8 ? plugin_fullname + '       ' : plugin_fullname)
@@ -242,7 +242,7 @@ function make_delegate (instance, plugin, options) {
   }
   Logging.makelogfuncs(delegate)
 
-  delegate.die = options.makedie(delegate, {type: 'plugin', plugin: plugin.name})
+  delegate.die = Common.makedie(delegate, { type: 'plugin', plugin: plugin.name })
 
   var actmetalist = []
