+++ lib/transport.js
@@ -43,7 +43,7 @@ exports.listen = function (callpoint) {
   }
 }
 
-exports.client = function (callpoint, makedie) {
+exports.client = function (callpoint) {
   return function api_client () {
     var self = this
 
@@ -75,35 +75,27 @@ exports.client = function (callpoint, makedie) {
       return _.isString(pin) ? Jsonic(pin) : pin
     })
 
-    var sd = Plugins.make_delegate(
-      self,
-      {
-        name: 'client$',
-        tag: void 0
-      },
-      {
-        makedie: makedie
-      })
+    var sd = Plugins.make_delegate(self, { name: 'client$', tag: void 0 })
 
-    var transport_client = function transport_client ( msg, done ) {
-      if ( msg.local$ ) {
-        this.prior( msg, done )
+    var transport_client = function transport_client (msg, done) {
+      if (msg.local$) {
+        this.prior(msg, done )
       }
       else {
-        sendclient.send.call( this, msg, done )
+        sendclient.send.call(this, msg, done )
       }
     }
 
     transport_client.id = config.id
 
-    if ( config.handle ) {
+    if (config.handle) {
       transport_client.handle = config.handle
     }
 
-    _.each( pins, function (pin) {
+    _.each(pins, function (pin) {
       pin.client$ = true
 
-      sd.add( pin, transport_client)
+      sd.add(pin, transport_client)
     })
 
     // Create client.
@@ -112,17 +104,17 @@ exports.client = function (callpoint, makedie) {
       { config: config, gate$: true },
       function (err, liveclient) {
         if (err) {
-          return sd.die( internals.error(err, 'transport_client', config) )
+          return sd.die(internals.error(err, 'transport_client', config) )
         }
-        if ( null === liveclient ) {
-          return sd.die( internals.error('transport_client_null',
+        if (liveclient === null) {
+          return sd.die(internals.error('transport_client_null',
                                          Common.clean(config)))
         }
 
         // Process any messages waiting for this client,
         // before bringing client online.
         function sendnext () {
-          if (sendqueue.length === 0) {
+          if (!sendqueue.length) {
             sendclient = liveclient
             sd.log.debug('client', 'sendqueue-clear', config)
           }
